# Split out libc tests into one enclave binary per test