… cmake/add_enclave.cmake
… tests/libc/CMakeLists.txt
… tests/libc/enc/enc.c
… tests/libc/enc/tests.c.in