Previously, the libc tests were all run as part of a single enclave binary. Additionally, they would fail after a single test, meaning other tests were not run after the first failure.

Update the libc tests so that each test in libc-test is run as a separate test in cmake.

Note that this does have a perf impact (both at build time and test running time) but I still think it's the right thing to do.

Fixes #2487