@radhikaj 2020-01-31
@jhand2 When a developer executes ctest after building the SDK, how much longer does it take after you split tests up?