@jhand2 2020-01-31
> @jhand2 When a developer executes ctest after building the SDK, how much longer does it take after you split tests up?

I ran just one pass for each of these benchmarks so there could be some variation but probably not much.

In simulation mode:
**Before**: 129.90 seconds
**After**: 149.96 seconds

On SGX hardware:
**Before**: 296.74
**After**: 308.41 seconds

So about 20 seconds slower on both. Not great (but you feel the difference more in simulation mode). But before this change debugging a libc test failure is very challenging