@jhand2 2020-02-01
@radhikaj @CodeMonkeyLeet @anakrish Do you see the increase in elapsed time to run tests (+20 seconds) as to great to justify this change? Or do you think the debugging benefit is worth it?

For reference, each libc test takes 0.05 seconds in simulation mode, 0.1 seconds on SGX. It just adds so much time because there are so many of them.