@anakrish 2020-02-01
@jhand2 Please do follow up with @mikbras. If memory serves right, libC tests used to take a lot of time due to there being a large number of tests, and Mike optimized the running time by combining them.