@jhand2 2020-02-01
> Regarding the question of the additional runtime overhead, have we considered making a build/ctest toggle? There's precedence here for this in how we choose to run a subset of libcxx tests in the default case and a nightly run of CI/CD runs the full set. Similarly, if the expectation is that these usually pass and do not need devs to investigate failures in their inner loop regularly, it's useful to default to the previous behavior where it's treated as a single ctest, with the option to split them into separate ctests to enable deeper debugging.

Good point, a build toggle is probably the right way to go.