@jhand2 2020-02-03
I have added a compile option. Now if you run cmake with `-DOE_SEPARATE_LIBC_TESTS`, each libc test will be run as a separate enclave.

bors try