@jhand2 2020-02-04
Fixed merge conflict

bors try