@jhand2 2020-02-04
> Thanks for making this! LGTM with only minor comments.

Sorry, I took a while to close this. https://github.com/openenclave/openenclave/pull/2501 replaces this PR